gps/requirements.txt
pyserial
pynmea2
